renamed Icon component to IconLink
